+++ src/routes/index.routes.ts
@@ -1,13 +1,13 @@
 import { Request, Response, Router } from "express";
-import UsersController from "../controllers/Users.controller";
-import UsersListProductsController from "../controllers/UsersListProducts.controller";
+import UsersController from "../controllers/users.controller";
+import ListProductsController from "../controllers/listProducts.controller";
 import { createValidator } from "express-joi-validation";
-import { bodySchemaLogin, bodySchemaRegister, bodySchemaUpdate, paramsSchemaUpdate } from "../middlewares/usersValidators";
+import { bodySchemaLogin, bodySchemaRegister, bodySchemaUpdate } from "../middlewares/usersValidators";
 import { authUser } from "../auth";
-import { paramsSchemaProducts } from "../middlewares/usersListValidator";
+import { paramsSchemaProducts } from "../middlewares/listValidator";
 
 const usersController = new UsersController();
-const usersListProductsController = new UsersListProductsController();
+const listProductsController = new ListProductsController();
 
 const router = Router();
 const validator = createValidator({
@@ -21,12 +21,7 @@ router.post('/login', validator.body(bodySchemaLogin), usersController.find)
 router.post('/register', validator.body(bodySchemaRegister), usersController.store)
 router.post('/user/:id', authUser, validator.body(bodySchemaUpdate), usersController.update)
 // Users List Route
-router.get(
-  '/products/:id',
-  authUser,
-  validator.params(paramsSchemaProducts),
-  usersListProductsController.findProducts
-)
+router.get('/products/:id', authUser, validator.params(paramsSchemaProducts), listProductsController.findProducts);
 
 
 export default router;
